requirements.txt
openai
python-dotenv
pydantic
streamlit
langchain
beautifulsoup4
selenium